# Remove unnecessary decimal rescaling - memory usage optimization