--- a/decimal_test.go
+++ b/decimal_test.go
@@ -9,7 +9,6 @@ import (
 	"math/big"
 	"math/rand"
 	"reflect"
-	"sort"
 	"strconv"
 	"strings"
 	"testing"
@@ -198,47 +197,6 @@ func TestNewFromFloat32Quick(t *testing.T) {
 	}, &quick.Config{})
 	if err != nil {
 		t.Error(err)
-	}
-}
-
-func BenchmarkNewFromFloatWithExponent(b *testing.B) {
-	rng := rand.New(rand.NewSource(0xdead1337))
-	in := make([]float64, b.N)
-	for i := range in {
-		in[i] = rng.NormFloat64() * 10e20
-	}
-	b.ReportAllocs()
-	b.StartTimer()
-	for i := 0; i < b.N; i++ {
-		in := rng.NormFloat64() * 10e20
-		_ = NewFromFloatWithExponent(in, math.MinInt32)
-	}
-}
-
-func BenchmarkNewFromFloat(b *testing.B) {
-	rng := rand.New(rand.NewSource(0xdead1337))
-	in := make([]float64, b.N)
-	for i := range in {
-		in[i] = rng.NormFloat64() * 10e20
-	}
-	b.ReportAllocs()
-	b.StartTimer()
-	for i := 0; i < b.N; i++ {
-		_ = NewFromFloat(in[i])
-	}
-}
-
-func BenchmarkNewFromStringFloat(b *testing.B) {
-	rng := rand.New(rand.NewSource(0xdead1337))
-	in := make([]float64, b.N)
-	for i := range in {
-		in[i] = rng.NormFloat64() * 10e20
-	}
-	b.ReportAllocs()
-	b.StartTimer()
-	for i := 0; i < b.N; i++ {
-		in := strconv.FormatFloat(in[i], 'f', -1, 64)
-		_, _ = NewFromString(in)
 	}
 }
 
@@ -859,22 +817,6 @@ func TestDecimal_Floor(t *testing.T) {
 	for _, test := range testsWithDecimals {
 		expected, _ := NewFromString(test.expected)
 		assertFloor(test.input, expected)
-	}
-}
-
-func Benchmark_FloorFast(b *testing.B) {
-	input := New(200, 2)
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		input.Floor()
-	}
-}
-
-func Benchmark_FloorRegular(b *testing.B) {
-	input := New(200, -2)
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		input.Floor()
 	}
 }
 
@@ -1525,44 +1467,6 @@ func (d Decimal) DivOld(d2 Decimal, prec int) Decimal {
 	return ret
 }
 
-func Benchmark_DivideOriginal(b *testing.B) {
-	tcs := createDivTestCases()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		for _, tc := range tcs {
-			d := tc.d
-			if sign(tc.d2) == 0 {
-				continue
-			}
-			d2 := tc.d2
-			prec := tc.prec
-			a := d.DivOld(d2, int(prec))
-			if sign(a) > 2 {
-				panic("dummy panic")
-			}
-		}
-	}
-}
-
-func Benchmark_DivideNew(b *testing.B) {
-	tcs := createDivTestCases()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		for _, tc := range tcs {
-			d := tc.d
-			if sign(tc.d2) == 0 {
-				continue
-			}
-			d2 := tc.d2
-			prec := tc.prec
-			a := d.DivRound(d2, prec)
-			if sign(a) > 2 {
-				panic("dummy panic")
-			}
-		}
-	}
-}
-
 func sign(d Decimal) int {
 	return d.value.Sign()
 }
@@ -1717,26 +1621,6 @@ func TestDecimal_RoundCash_Panic(t *testing.T) {
 	}()
 	d, _ := NewFromString("1")
 	d.RoundCash(231)
-}
-
-func BenchmarkDecimal_RoundCash_Five(b *testing.B) {
-	const want = "3.50"
-	for i := 0; i < b.N; i++ {
-		val := New(3478, -3)
-		if have := val.StringFixedCash(5); have != want {
-			b.Fatalf("\nHave: %q\nWant: %q", have, want)
-		}
-	}
-}
-
-func BenchmarkDecimal_RoundCash_Fifteen(b *testing.B) {
-	const want = "6.30"
-	for i := 0; i < b.N; i++ {
-		val := New(635, -2)
-		if have := val.StringFixedCash(15); have != want {
-			b.Fatalf("\nHave: %q\nWant: %q", have, want)
-		}
-	}
 }
 
 func TestDecimal_Mod(t *testing.T) {
@@ -2190,22 +2074,6 @@ func TestDecimal_Coefficient(t *testing.T) {
 	co.Set(big.NewInt(0))
 	if d.IntPart() != 123 {
 		t.Error("Modifying coefficient modified Decimal; Got:", d)
-	}
-}
-
-type DecimalSlice []Decimal
-
-func (p DecimalSlice) Len() int           { return len(p) }
-func (p DecimalSlice) Swap(i, j int)      { p[i], p[j] = p[j], p[i] }
-func (p DecimalSlice) Less(i, j int) bool { return p[i].Cmp(p[j]) < 0 }
-func Benchmark_Cmp(b *testing.B) {
-	decimals := DecimalSlice([]Decimal{})
-	for i := 0; i < 1000000; i++ {
-		decimals = append(decimals, New(int64(i), 0))
-	}
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		sort.Sort(decimals)
 	}
 }