@mwoss 2019-12-29
Tests failing on Go1.2.2, I'll fix it tomorrow.