@mwoss 2019-12-31
Removed benchmarks using big.NewFromFloat (my bad adding them). I'll leave comparison table in the main commnet.