@njason 2020-01-05
`RescalePair` looks good! I'm hesitant to approve with the formatting changes added since it is difficult to see what code is new in the `Sin` `Cos` and `Tan` functions. I made a separate PR to just format the code #162 . Once it is merged it should clean up this PR up a lot.