@mwoss 2020-01-05
Oh yea, my bad. I think Golnand format code automatically.